Release/desarrollo

+++ src/main/java/controlador/ControladorCategorias.java
@@ -11,6 +11,8 @@ import javax.servlet.ServletException;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
+import javax.servlet.http.HttpSession;
+import javax.swing.JOptionPane;
 import modelo.DAO.CategoriaDAO;
 import modelo.POJO.Categoria;
 
@@ -22,6 +24,7 @@ public class ControladorCategorias extends HttpServlet {
 
     CategoriaDAO dao = new CategoriaDAO();
     Categoria cate = new Categoria();
+
     /**
      * Processes requests for both HTTP <code>GET</code> and <code>POST</code>
      * methods.
@@ -39,7 +42,7 @@ public class ControladorCategorias extends HttpServlet {
             out.println("<!DOCTYPE html>");
             out.println("<html>");
             out.println("<head>");
-            out.println("<title>Servlet ControladorCategorias</title>");            
+            out.println("<title>Servlet ControladorCategorias</title>");
             out.println("</head>");
             out.println("<body>");
             out.println("<h1>Servlet ControladorCategorias at " + request.getContextPath() + "</h1>");
@@ -60,6 +63,14 @@ public class ControladorCategorias extends HttpServlet {
     @Override
     protected void doGet(HttpServletRequest request, HttpServletResponse response)
             throws ServletException, IOException {
+        int id_categorias = Integer.parseInt(request.getParameter("idCategoria"));
+        
+        HttpSession session =request.getSession();
+        session.setAttribute("categoria_actual", id_categorias);
+        
+        response.sendRedirect("bolsillos.jsp");
+    
+        
     }
 
     /**
@@ -73,7 +84,14 @@ public class ControladorCategorias extends HttpServlet {
     @Override
     protected void doPost(HttpServletRequest request, HttpServletResponse response)
             throws ServletException, IOException {
-              String accion = request.getParameter("accion");
+        int id_categorias = Integer.parseInt(request.getParameter("idCategoria"));
+        
+        HttpSession session =request.getSession();
+        session.setAttribute("categoria_actual", id_categorias);
+        
+        response.sendRedirect("bolsillos.jsp");
+        
+        String accion = request.getParameter("accion");
         switch (accion) {
             case "Listar":
 
@@ -92,59 +110,55 @@ public class ControladorCategorias extends HttpServlet {
             case "Actualizar":
                 String id1 = request.getParameter("txtid");
                 String nom1 = request.getParameter("txtnom");
-                String presupuesto1 = request.getParameter("txtpresupuesto");
-                String gasto1 = request.getParameter("txtgasto");
 
                 int idint = Integer.parseInt(id1);
-                long presupuestlolong = Long.parseLong(presupuesto1);
-                long gastolong = Long.parseLong(gasto1);
-                
+
                 cate.setId_categoria(idint);
                 cate.setNombre_categoria(nom1);
-                cate.setPresupuesto_categoria(presupuestlolong);
-                cate.setGasto_categoria(gastolong);
-                
+
                 dao.actualizar(cate);
                 request.getRequestDispatcher("ControladorCategorias?accion=Listar").forward(request, response);
                 break;
             case "Delete":
                 String id2 = request.getParameter("id");
-                dao.delete(id2);
+                int num_bolsillos = dao.num_bolsillos(id2);
+                if (num_bolsillos > 0) {
+
+                    JOptionPane.showMessageDialog(null, "No se puede borrar debido a bolsillos asociados");
+                    System.out.println("No se puede borrar debido a bolsillos asociados");
+
+                } else {
+                    dao.delete(id2);
+                }
+
                 request.getRequestDispatcher("ControladorCategorias?accion=Listar").forward(request, response);
                 break;
 
             case "Nuevo":
                 request.getRequestDispatcher("agregarCategorias.jsp").forward(request, response);
                 break;
-                
-                
+
             case "submit":
-                
+
                 String id = request.getParameter("txtid");
                 String nombres = request.getParameter("txtnombres");
-                String presupuesto = request.getParameter("txtpresupuesto");
-                String gasto = request.getParameter("txtgasto");
-                
+
                 int id3 = Integer.parseInt(id);
-                long presupuesto3 = Long.parseLong(presupuesto);
-                long gasto3 = Long.parseLong(gasto);
-                
+
                 cate.setId_categoria(id3);
                 cate.setNombre_categoria(nombres);
-                cate.setPresupuesto_categoria(presupuesto3);
-                cate.setGasto_categoria(gasto3);
-                
+
                 dao.agregar(cate);
                 request.getRequestDispatcher("ControladorCategorias?accion=Listar").forward(request, response);
-                
-                    
-                    break;
+
+                break;
             default:
                 throw new AssertionError();
-        }           
-
+        }
 
     }
+
+   
 
     /**
      * Returns a short description of the servlet.
